requirements.txt
torch>=1.3
numpy>=1.16
pillow